refactor(analysis-types): extract complexity file DTOs
